Extract CPU debug module
