+++ src/cpu/mod.rs
@@ -1042,6 +1042,7 @@ impl Cpu {
     0
   }
 
+  #[cfg(feature = "terminal_debug")]
   pub fn disassemble(&mut self, start: u16, end: u16) -> HashMap<u16, String> {
     let mut addr = start as u32;
     let mut map: HashMap<u16, String> = HashMap::new();
@@ -1135,6 +1136,7 @@ impl Cpu {
     map
   }
 
+  #[cfg(feature = "terminal_debug")]
   fn extract_addr_16(&mut self, mut addr: u32) -> (u16, u16) {
     let lo_byte = self.bus_read_u8(addr.try_into().unwrap());
     addr += 1;
